refactor: 레포지토리 구조 개선
역할에 맞게 세분화
