+++ src/common/di/container.js
@@ -9,7 +9,10 @@ import { prisma } from '#db/prisma.js';
 import {
   UserRepository,
   ChallengeRepository,
+  ChallengeParticipantRepository,
+  DraftRepository,
   SubmissionRepository,
+  SubmissionLikeRepository,
   NotificationRepository,
 } from '#repository';
 import {
@@ -44,7 +47,16 @@ export const createContainer = () => {
     challengeRepository: asClass(ChallengeRepository, {
       lifetime: Lifetime.SINGLETON,
     }),
+    ChallengeParticipantRepository: asClass(ChallengeParticipantRepository, {
+      lifetime: Lifetime.SINGLETON,
+    }),
+    draftRepository: asClass(DraftRepository, {
+      lifetime: Lifetime.SINGLETON,
+    }),
     submissionRepository: asClass(SubmissionRepository, {
+      lifetime: Lifetime.SINGLETON,
+    }),
+    submissionLikeRepository: asClass(SubmissionLikeRepository, {
       lifetime: Lifetime.SINGLETON,
     }),
     notificationRepository: asClass(NotificationRepository, {
