+++ src/repository/index.js
@@ -2,3 +2,6 @@ export * from './user.repository.js';
 export * from './challenge.repository.js';
 export * from './submission.repository.js';
 export * from './notification.repository.js';
+export * from './challengeParticipant.repository.js';
+export * from './draft.repository.js';
+export * from './submissionLike.repository.js';
